Report: add env and allow skipping Linden setup

+++ report/build.sh
@@ -7,9 +7,11 @@ ARGS=$(jq -r '."tinymist.typstExtraArgs" | join(" ")' .vscode/settings.json)
 LINDEN_REF=$(echo "$ARGS" | sed -n 's/.*LINDEN_REF=\([^ ]*\).*/\1/p')
 
 # Download Linden at the specified ref. This will be used by typst during compilation.
-rm -rf Linden || true
-git clone https://github.com/epfl-systemf/Linden Linden
-git -C Linden reset --hard "$LINDEN_REF"
+if [[ ! "$*" =~ --skip-setup ]]; then
+	rm -rf Linden || true
+	git clone https://github.com/epfl-systemf/Linden Linden
+	git -C Linden reset --hard "$LINDEN_REF"
+fi
 
 typst compile thesis.typ thesis.pdf $ARGS
